setup initial api documentation e update settings.js file
