+++ src/scripts/theatre_main.js
@@ -1,7 +1,9 @@
+import API from "./API/api.js";
 import { KHelpers } from "./KHelpers.js";
 import { Theatre } from "./Theatre.js";
 import CONSTANTS from "./constants/constants.js";
-import { debug, log } from "./lib/lib.js";
+import { debug, log, warn } from "./lib/lib.js";
+import { registerKeybindings } from "./settings.js";
 
 /**
  * Concat helper
@@ -24,7 +26,9 @@ Handlebars.registerHelper("resprop", function (propPath, hash) {
  * Hook in on Actorsheet's Header buttons + context menus
  */
 Hooks.on("getActorSheetHeaderButtons", (app, buttons) => {
-  if (!game.user.isGM && game.settings.get("theatre", "gmOnly")) return;
+  if (!game.user.isGM && game.settings.get("theatre", "gmOnly")) {
+    return;
+  }
   const removeLabelSheetHeader = game.settings.get(CONSTANTS.MODULE_ID, "removeLabelSheetHeader");
 
   let theatreButtons = [];
@@ -59,8 +63,9 @@ Hooks.on("getActorSheetHeaderButtons", (app, buttons) => {
  */
 Hooks.on("sidebarCollapse", function (a, collapsed) {
   // If theatre isn't even ready, then just no
-  if (!Theatre.instance) return;
-
+  if (!Theatre.instance) {
+    return;
+  }
   debug("collapse? : ", a, collapsed);
   let sideBar = document.getElementById("sidebar");
   let primeBar = document.getElementById("theatre-prime-bar");
@@ -83,8 +88,9 @@ Hooks.on("sidebarCollapse", function (a, collapsed) {
   }
   Theatre.instance.theatreEmoteMenu.style.top = `${Theatre.instance.theatreControls.offsetTop - 410}px`;
 
-  if (Theatre.instance.reorderTOId) window.clearTimeout(Theatre.instance.reorderTOId);
-
+  if (Theatre.instance.reorderTOId) {
+    window.clearTimeout(Theatre.instance.reorderTOId);
+  }
   Theatre.instance.reorderTOId = window.setTimeout(() => {
     Theatre.reorderInserts();
     Theatre.instance.reorderTOId = null;
@@ -96,8 +102,9 @@ Hooks.on("sidebarCollapse", function (a, collapsed) {
  */
 Hooks.on("createCombat", function () {
   // If theatre isn't even ready, then just no
-  if (!Theatre.instance) return;
-
+  if (!Theatre.instance) {
+    return;
+  }
   if (!!game.combats.active && game.combats.active.round == 0 && Theatre.instance.isSuppressed) {
     debug("COMBAT CREATED");
     // if suppressed, change opacity to 0.05
@@ -113,8 +120,9 @@ Hooks.on("createCombat", function () {
  */
 Hooks.on("deleteCombat", function () {
   // If theatre isn't even ready, then just no
-  if (!Theatre.instance) return;
-
+  if (!Theatre.instance) {
+    return;
+  }
   if (!game.combats.active && Theatre.instance.isSuppressed) {
     debug("COMBAT DELETED");
     // if suppressed, change opacity to 0.25
@@ -228,8 +236,9 @@ Hooks.on("createChatMessage", function (chatEntity, _, userId) {
   let theatreId = null;
 
   // If theatre isn't even ready, then just no
-  if (!Theatre.instance) return;
-
+  if (!Theatre.instance) {
+    return;
+  }
   if (Theatre.instance.usersTyping[userId]) {
     theatreId = Theatre.instance.usersTyping[userId].theatreId;
     Theatre.instance.removeUserTyping(userId);
@@ -246,9 +255,9 @@ Hooks.on("createChatMessage", function (chatEntity, _, userId) {
     //|| Object.keys(chatData.speaker).length == 0
     chatData.content.match(/@[a-zA-Z0-9]+\[[a-zA-Z0-9]+\]/) ||
     chatData.content.match(/\<div.*\>[\s\S]*\<\/div\>/)
-  )
+  ) {
     return;
-
+  }
   let textBox = Theatre.instance.getTextBoxById(theatreId);
   let insert = Theatre.instance.getInsertById(theatreId);
   let charSpans = [];
@@ -396,7 +405,9 @@ Hooks.on("createChatMessage", function (chatEntity, _, userId) {
     );
 
     // auto decay?
-    if (insert && insert.decayTOId) window.clearTimeout(insert.decayTOId);
+    if (insert && insert.decayTOId) {
+      window.clearTimeout(insert.decayTOId);
+    }
     if (insert && Theatre.instance.settings.autoDecay) {
       insert.decayTOId = window.setTimeout(
         (imgId) => {
@@ -411,13 +422,16 @@ Hooks.on("createChatMessage", function (chatEntity, _, userId) {
 });
 
 Hooks.on("renderChatMessage", function (ChatMessage, html, data) {
-  if (Theatre.instance.settings.ignoreMessagesToChat && ChatMessage.flags?.[CONSTANTS.MODULE_ID]?.theatreMessage)
+  if (Theatre.instance.settings.ignoreMessagesToChat && ChatMessage.flags?.[CONSTANTS.MODULE_ID]?.theatreMessage) {
     html[0].style.display = "none";
+  }
   return true;
 });
 
 Hooks.on("renderChatLog", function (app, html, data) {
-  if (data.cssId === "chat-popout") return;
+  if (data.cssId === "chat-popout") {
+    return;
+  }
   theatre.initialize();
   if (!window.Theatre) {
     window.Theatre = Theatre;
@@ -434,8 +448,9 @@ Hooks.on("renderChatLog", function (app, html, data) {
  * Add to stage button on ActorDirectory Sidebar
  */
 Hooks.on("getActorDirectoryEntryContext", async (html, options) => {
-  if (!game.user.isGM && game.settings.get("theatre", "gmOnly")) return;
-
+  if (!game.user.isGM && game.settings.get("theatre", "gmOnly")) {
+    return;
+  }
   const getActorData = (target) => {
     return game.actors.get(target.data("documentId"));
   };
@@ -459,309 +474,29 @@ Hooks.on("getActorDirectoryEntryContext", async (html, options) => {
 });
 
 // Fixed global singleton/global object
-var theatre = null;
+let theatre = null;
 Hooks.once("setup", () => {
   theatre = new Theatre();
 
+  game.modules.get(CONSTANTS.MODULE_ID).api = API;
+
   // module keybinds
-  game.keybindings.register("theatre", "unfocusTextArea", {
-    name: "Theatre.UI.Keybinds.unfocusTextArea",
-    hint: "",
-    editable: [
-      {
-        key: "Escape",
-      },
-    ],
-    onDown: () => {
-      if (document.activeElement === document.getElementById("chat-message")) {
-        event.preventDefault();
-        event.stopPropagation();
-        document.getElementById("chat-message").blur();
-      }
-    },
-    restricted: false,
-  });
-
-  game.keybindings.register("theatre", "addOwnedToStage", {
-    name: "Theatre.UI.Keybinds.addOwnedToStage",
-    hint: "",
-    editable: [
-      {
-        key: "Enter",
-        modifiers: ["Alt"],
-      },
-    ],
-    onDown: () => {
-      const ownedActors = game.actors.filter((a) => a.permission === 3);
-      const ownedTokens = ownedActors.map((a) => a.getActiveTokens());
-      for (const tokenArray of ownedTokens) tokenArray.forEach((t) => Theatre.addToNavBar(t.actor));
-    },
-    restricted: false,
-  });
-
-  game.keybindings.register("theatre", "addSelectedToStage", {
-    name: "Theatre.UI.Keybinds.addSelectedToStage",
-    hint: "",
-    editable: [
-      {
-        key: "Enter",
-        modifiers: ["Shift"],
-      },
-    ],
-    onDown: () => {
-      for (const tkn of canvas.tokens.controlled) Theatre.addToNavBar(tkn.actor);
-    },
-    restricted: true,
-  });
-
-  game.keybindings.register("theatre", `removeSelectedFromStage`, {
-    name: "Theatre.UI.Keybinds.removeSelectedFromStage",
-    hint: "",
-    editable: [],
-    onDown: (context) => {
-      for (const tkn of canvas.tokens.controlled) Theatre.removeFromNavBar(tkn.actor);
-    },
-    restricted: true,
-  });
-
-  game.keybindings.register("theatre", "narratorMode", {
-    name: "Theatre.UI.Keybinds.narratorMode",
-    hint: "",
-    editable: [
-      {
-        key: "KeyN",
-        modifiers: ["Alt"],
-      },
-    ],
-    onDown: () => {
-      const narratorButton = $(document).find(`div.theatre-icon-narrator`).closest(`div.theatre-control-btn`);
-      if (KHelpers.hasClass(narratorButton[0], "theatre-control-nav-bar-item-speakingas"))
-        Theatre.instance.toggleNarratorBar(false);
-      else Theatre.instance.toggleNarratorBar(true);
-
-      document.getElementById("chat-message").blur();
-    },
-    restricted: true,
-  });
-
-  game.keybindings.register("theatre", "flipPortrait", {
-    name: "Theatre.UI.Keybinds.flipPortrait",
-    hint: "",
-    editable: [
-      {
-        key: "KeyR",
-        modifiers: ["Alt"],
-      },
-    ],
-    onDown: () => {
-      if (Theatre.instance.speakingAs) Theatre.instance.mirrorInsertById(Theatre.instance.speakingAs);
-    },
-    restricted: false,
-  });
-
-  game.keybindings.register("theatre", "nudgePortraitLeft", {
-    name: "Theatre.UI.Keybinds.nudgePortraitLeft",
-    hint: "",
-    editable: [
-      {
-        key: "KeyZ",
-        modifiers: ["Alt"],
-      },
-    ],
-    onDown: () => {
-      const imgId = Theatre.instance.speakingAs;
-      if (!imgId) return;
-
-      const insert = Theatre.instance.portraitDocks.find((p) => p.imgId === imgId);
-      const oleft = insert.portraitContainer.x,
-        otop = insert.portraitContainer.y;
-      const tweenId = "portraitMove";
-      const tween = TweenMax.to(insert.portraitContainer, 0.5, {
-        pixi: { x: oleft - 50, y: otop },
-        ease: Power3.easeOut,
-        onComplete: function (ctx, imgId, tweenId) {
-          // decrement the rendering accumulator
-          ctx._removeDockTween(imgId, this, tweenId);
-          // remove our own reference from the dockContainer tweens
-        },
-        onCompleteParams: [Theatre.instance, insert.imgId, tweenId],
-      });
-      Theatre.instance._addDockTween(insert.imgId, tween, tweenId);
-
-      // send sceneEvent
-      Theatre.instance._sendSceneEvent("positionupdate", {
-        insertid: insert.imgId,
-        position: { x: oleft - 50, y: otop, mirror: insert.mirrored },
-      });
-    },
-    restricted: false,
-  });
-
-  game.keybindings.register("theatre", "nudgePortraitRight", {
-    name: "Theatre.UI.Keybinds.nudgePortraitRight",
-    hint: "",
-    editable: [
-      {
-        key: "KeyC",
-        modifiers: ["Alt"],
-      },
-    ],
-    onDown: () => {
-      const imgId = Theatre.instance.speakingAs;
-      if (!imgId) return;
-
-      const insert = Theatre.instance.portraitDocks.find((p) => p.imgId === imgId);
-      const oleft = insert.portraitContainer.x,
-        otop = insert.portraitContainer.y;
-      const tweenId = "portraitMove";
-      const tween = TweenMax.to(insert.portraitContainer, 0.5, {
-        pixi: { x: oleft + 50, y: otop },
-        ease: Power3.easeOut,
-        onComplete: function (ctx, imgId, tweenId) {
-          // decrement the rendering accumulator
-          ctx._removeDockTween(imgId, this, tweenId);
-          // remove our own reference from the dockContainer tweens
-        },
-        onCompleteParams: [Theatre.instance, insert.imgId, tweenId],
-      });
-      Theatre.instance._addDockTween(insert.imgId, tween, tweenId);
-
-      // send sceneEvent
-      Theatre.instance._sendSceneEvent("positionupdate", {
-        insertid: insert.imgId,
-        position: { x: oleft + 50, y: otop, mirror: insert.mirrored },
-      });
-    },
-    restricted: false,
-  });
-
-  game.keybindings.register("theatre", "nudgePortraitUp", {
-    name: "Theatre.UI.Keybinds.nudgePortraitUp",
-    hint: "",
-    editable: [
-      {
-        key: "KeyS",
-        modifiers: ["Alt"],
-      },
-    ],
-    onDown: () => {
-      const imgId = Theatre.instance.speakingAs;
-      if (!imgId) return;
-
-      const insert = Theatre.instance.portraitDocks.find((p) => p.imgId === imgId);
-      const oleft = insert.portraitContainer.x,
-        otop = insert.portraitContainer.y;
-      const tweenId = "portraitMove";
-      const tween = TweenMax.to(insert.portraitContainer, 0.5, {
-        pixi: { x: oleft, y: otop - 50 },
-        ease: Power3.easeOut,
-        onComplete: function (ctx, imgId, tweenId) {
-          // decrement the rendering accumulator
-          ctx._removeDockTween(imgId, this, tweenId);
-          // remove our own reference from the dockContainer tweens
-        },
-        onCompleteParams: [Theatre.instance, insert.imgId, tweenId],
-      });
-      Theatre.instance._addDockTween(insert.imgId, tween, tweenId);
-
-      // send sceneEvent
-      Theatre.instance._sendSceneEvent("positionupdate", {
-        insertid: insert.imgId,
-        position: { x: oleft, y: otop - 50, mirror: insert.mirrored },
-      });
-    },
-    restricted: false,
-  });
-
-  game.keybindings.register("theatre", "nudgePortraitDown", {
-    name: "Theatre.UI.Keybinds.nudgePortraitDown",
-    hint: "",
-    editable: [
-      {
-        key: "KeyX",
-        modifiers: ["Alt"],
-      },
-    ],
-    onDown: () => {
-      const imgId = Theatre.instance.speakingAs;
-      if (!imgId) return;
-
-      const insert = Theatre.instance.portraitDocks.find((p) => p.imgId === imgId);
-      const oleft = insert.portraitContainer.x,
-        otop = insert.portraitContainer.y;
-      const tweenId = "portraitMove";
-      const tween = TweenMax.to(insert.portraitContainer, 0.5, {
-        pixi: { x: oleft, y: otop + 50 },
-        ease: Power3.easeOut,
-        onComplete: function (ctx, imgId, tweenId) {
-          // decrement the rendering accumulator
-          ctx._removeDockTween(imgId, this, tweenId);
-          // remove our own reference from the dockContainer tweens
-        },
-        onCompleteParams: [Theatre.instance, insert.imgId, tweenId],
-      });
-      Theatre.instance._addDockTween(insert.imgId, tween, tweenId);
-
-      // send sceneEvent
-      Theatre.instance._sendSceneEvent("positionupdate", {
-        insertid: insert.imgId,
-        position: { x: oleft, y: otop + 50, mirror: insert.mirrored },
-      });
-    },
-    restricted: false,
-  });
-
-  for (let i = 1; i < 11; i++) {
-    game.keybindings.register("theatre", `activateStaged${i}`, {
-      name: game.i18n.format(`Theatre.UI.Keybinds.activateStaged`, { number: i }),
-      hint: "",
-      editable: [
-        {
-          key: `Digit${i === 10 ? 0 : i}`,
-          modifiers: ["Control"],
-        },
-      ],
-      onDown: () => {
-        const ids = Object.keys(Theatre.instance.stage);
-        const id = ids[i - 1];
-        if (id) Theatre.instance.activateInsertById(id);
-
-        document.getElementById("chat-message").blur();
-      },
-      restricted: false,
-      reservedModifiers: ["Shift"],
-    });
-
-    game.keybindings.register("theatre", `removeStaged${i}`, {
-      name: game.i18n.format(`Theatre.UI.Keybinds.removeStaged`, { number: i }),
-      hint: "",
-      editable: [
-        {
-          key: `Digit${i === 10 ? 0 : i}`,
-          modifiers: ["Control", "Alt"],
-        },
-      ],
-      onDown: () => {
-        const ids = Object.keys(Theatre.instance.stage);
-        const id = ids[i - 1];
-        if (id) Theatre.instance.removeInsertById(id);
-      },
-      restricted: true,
-    });
-  }
+  registerKeybindings();
 });
 
 /**
  * Hide player list (and macro hotbar) when stage is active (and not suppressed)
  */
 Hooks.on("theatreDockActive", (insertCount) => {
-  if (!insertCount) return;
-
+  if (!insertCount) {
+    return;
+  }
   // The "MyTab" module inserts another element with id "pause". Use querySelectorAll to make sure we catch both
   document.querySelectorAll("#pause").forEach((ele) => KHelpers.addClass(ele, "theatre-centered"));
 
-  if (!game.settings.get(CONSTANTS.MODULE_ID, "autoHideBottom")) return;
-
+  if (!game.settings.get(CONSTANTS.MODULE_ID, "autoHideBottom")) {
+    return;
+  }
   if (!theatre.isSuppressed) {
     $("#players").addClass("theatre-invisible");
     $("#hotbar").addClass("theatre-invisible");
@@ -772,28 +507,47 @@ Hooks.on("theatreDockActive", (insertCount) => {
  * If Argon is active, wrap CombatHudCanvasElement#toggleMacroPlayers to prevent playesr list and macro hotbar from being shown
  */
 Hooks.once("ready", () => {
-  if (!game.settings.get(CONSTANTS.MODULE_ID, "autoHideBottom")) return;
-  if (!game.modules.get("enhancedcombathud")?.active) return;
-
+  if (!game.modules.get("socketlib")?.active && game.user?.isGM) {
+    let word = "install and activate";
+    if (game.modules.get("socketlib")) word = "activate";
+    warn(`It is recommended to intall the 'socketlib' module. Please ${word} it.`);
+  }
+  if (!game.settings.get(CONSTANTS.MODULE_ID, "autoHideBottom")) {
+    return;
+  }
+  if (!game.modules.get("enhancedcombathud")?.active) {
+    return;
+  }
   libWrapper.register(
     CONSTANTS.MODULE_ID,
     "CombatHudCanvasElement.prototype.toggleMacroPlayers",
     (wrapped, togg) => {
-      if (togg && theatre?.dockActive) return;
+      if (togg && theatre?.dockActive) {
+        return;
+      }
       return wrapped(togg);
     },
     "MIXED"
   );
 });
 
+Hooks.once("devModeReady", ({ registerPackageDebugFlag }) => {
+  registerPackageDebugFlag(CONSTANTS.MODULE_ID);
+});
+
 /**
  * Hide/show macro hotbar when stage is suppressed
  */
 Hooks.on("theatreSuppression", (suppressed) => {
-  if (!game.settings.get(CONSTANTS.MODULE_ID, "autoHideBottom")) return;
-  if (!game.settings.get(CONSTANTS.MODULE_ID, "suppressMacroHotbar")) return;
-  if (!theatre.dockActive) return;
-
+  if (!game.settings.get(CONSTANTS.MODULE_ID, "autoHideBottom")) {
+    return;
+  }
+  if (!game.settings.get(CONSTANTS.MODULE_ID, "suppressMacroHotbar")) {
+    return;
+  }
+  if (!theatre.dockActive) {
+    return;
+  }
   if (suppressed) {
     $("#players").removeClass("theatre-invisible");
     $("#hotbar").removeClass("theatre-invisible");
@@ -804,7 +558,9 @@ Hooks.on("theatreSuppression", (suppressed) => {
 });
 
 Hooks.on("renderPause", () => {
-  if (!theatre?.dockActive) return;
+  if (!theatre?.dockActive) {
+    return;
+  }
   // The "MyTab" module inserts another element with id "pause". Use querySelectorAll to make sure we catch both
   document.querySelectorAll("#pause").forEach((ele) => KHelpers.addClass(ele, "theatre-centered"));
 });
@@ -814,8 +570,9 @@ Hooks.on("renderPause", () => {
  */
 Hooks.on("updateActor", (actor, data) => {
   const insert = Theatre.instance.getInsertById(`theatre-${actor.id}`);
-  if (!insert) return;
-
+  if (!insert) {
+    return;
+  }
   insert.label.text = Theatre.getActorDisplayName(actor.id);
   Theatre.instance._renderTheatre(performance.now());
 });
@@ -829,7 +586,9 @@ Hooks.on("getSceneControlButtons", (controls) => {
       icon: "fas fa-theater-masks",
       toggle: true,
       active: false,
-      onClick: (toggle) => Theatre.instance.updateSuppression(toggle), // TODO Suppress theatre
+      onClick: (toggle) => {
+        Theatre.instance.updateSuppression(toggle); // TODO Suppress theatre
+      },
       visible: true,
     };
     const tokenControls = controls.find((group) => group.name === "token").tools;
