+++ wiki/api.md
@@ -1,0 +1,17 @@
+The api is reachable from the variable `game.modules.get('theatre').api` or from the socket library `socketLib` on the variable `game.modules.get('theatre').socket` if present and active.
+
+### The documentation can be out of sync with the API code checkout the code if you want to dig up [API](../src/scripts/API/api.js)
+
+You can find some javascript examples here **=> [macros](./macros/) <=**
+
+TODO
+
+#### Name o the method
+
+**Returns**: <code>Promise&lt;void&gt;</code> - Return nothing
+
+| Param | Type | Description | Note |
+| --- | --- | --- | --- |
+
+
+**Example basic**:
